service-workers/service-worker/installing.https.html

<!DOCTYPE html>
<title>ServiceWorker: navigator.serviceWorker.installing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>

const SCRIPT = 'resources/empty-worker.js';
const SCOPE = 'resources/blank.html';
let registration;

// "installing" is set
promise_test(async t => {
    let step = t.step_func.bind(t);
    let frame;

    await service_worker_unregister(t, SCOPE);
    frame = await with_iframe(SCOPE);
    registration =
        await navigator.serviceWorker.register(SCRIPT, {scope: SCOPE});
    const container = await frame.contentWindow.navigator.serviceWorker;
    assert_equals(container.controller, null);
    assert_equals(registration.active, null);
    assert_equals(registration.waiting, null);
    assert_equals(registration.installing.scriptURL, normalizeURL(SCRIPT));
    // FIXME: Add a test for a frame created after installation.
    // Should the existing frame ("frame") block activation?

    t.add_cleanup(() => frame.remove());
    await service_worker_unregister_and_done(t, SCOPE);
}, 'installing is set');

// Tests that The ServiceWorker objects returned from installing attribute getter
// that represent the same service worker are the same objects.
promise_test(async t => {
    registration =
        await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
    await registration.installing;
    const r = await navigator.serviceWorker.getRegistration(SCOPE);
    assert_equals(r.scriptURl, registration.scriptURl, 'getRegistration ' +
                  'should return the same object');
    t.add_cleanup(() => r.unregister());
}, 'The ServiceWorker objects returned from installing attribute getter that ' +
   'represent the same service worker are the same objects');
</script>
