third_party/WebKit/LayoutTests/platform/win/external/wpt/service-workers/service-worker/installing.https-expected.txt

This is a testharness.js-based test.
PASS installing is set 
Harness: the test ran to completion.
